requirements.txt
# AWS SDK
boto3==1.26.153
botocore==1.29.153

# AWS CDK Python bindings
aws-cdk-lib==2.87.0
constructs==10.2.69

# API and Lambda Function
fastapi==0.100.0
mangum==0.17.0
pydantic==2.0.2

# Testing
pytest==7.3.1
pytest-mock==3.11.1
moto==4.1.12

# Utilities
python-dotenv==1.0.0
structlog==23.1.0
pyyaml==6.0

# OpenAI Integration
openai==0.27.9

# External Services
twilio==8.5.0

# Type checking and linting
mypy==1.3.0
flake8==6.0.0
black==23.3.0
isort==5.12.0 